🏷️ Create type for timeline and timeline item
