--- src/lib/timeline/timelineItem/Type.ts
@@ -1,10 +1,0 @@
-export enum Type {
-	education = 'education',
-	internship = 'internship',
-	work = 'work',
-}
-
-export enum Position {
-	left = 'left',
-	right = 'right',
-}
